[Patchset 11] trace_processor: add code for parsing a trace
This CL utilises the previously introduced ProtoDecoder class to parse a
Trace into its important components. It doesn't do anything with the
parsed data yet but it lays the skeleton for writing to TraceStorage in
a subsequent CL.

Bug: 80416541

+++ src/trace_processor/trace_parser.cc
@@ -150,7 +150,7 @@ void TraceParser::ParseSchedSwitch(uint32_t cpu,
       case protos::SchedSwitchFtraceEvent::kNextCommFieldNumber:
         next_comm =
             std::string(reinterpret_cast<const char*>(fld.length_value.data),
-                        fld.length_value.length);
+                        static_cast<size_t>(fld.length_value.length));
         break;
       default:
         break;
